Changes to build on Mac using clang (merge from ferret/trunk)
git-svn-id: file:///home/users/tmap/svn/repos/ferret/branches/pyferret@24996 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ fer/ccr/get_edgmem.c
@@ -14,7 +14,7 @@
 #include "ferret.h"
 #include "FerMem.h"
 
-void FORTRAN(get_edgmem)( double *index, long *alen, int *status )
+void FORTRAN(get_edgmem)( int *index, long *alen, int *status )
 
 /*
   input  - index:  Ferret line number in which to store the array pointer
